fix checkout form translation
